requirements.txt
asgiref==3.8.1
certifi==2024.2.2
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
cloudinary==1.40.0
Django==5.0.6
django-channels==0.7.0
django-cloudinary-storage==0.3.0
django-environ==0.11.2
djangorestframework==3.15.1
idna==3.7
msgpack==1.0.8
oauthlib==3.2.2
pillow==10.3.0
PyJWT==2.8.0
redis==5.0.4
requests==2.32.1
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1